feat: edit model configuration on model version page
