+++ public/components/common/forms/model_configuration/index.ts
@@ -3,4 +3,4 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React from 'react';
+export * from './model_configuration';
