+++ public/components/register_model/model_configuration.tsx
@@ -3,100 +3,13 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-import React, { useState } from 'react';
-import {
-  EuiFormRow,
-  EuiCodeEditor,
-  EuiText,
-  EuiTextColor,
-  EuiCode,
-  EuiSpacer,
-  EuiButtonEmpty,
-} from '@elastic/eui';
-import { useController, useFormContext } from 'react-hook-form';
+import React from 'react';
+import { EuiText, EuiTextColor, EuiCode, EuiSpacer } from '@elastic/eui';
 
 import '../../ace-themes/sql_console.js';
-import type { ModelFileFormData, ModelUrlFormData } from './register_model.types';
-import { HelpFlyout } from './help_flyout';
-import { ErrorMessage } from './error_message';
-import { CUSTOM_FORM_ERROR_TYPES } from '../common/forms/form_constants';
-
-function validateConfigurationObject(value: string) {
-  try {
-    JSON.parse(value.trim());
-  } catch {
-    return 'JSON is invalid. Enter a valid JSON';
-  }
-  return true;
-}
-
-function validateModelType(value: string) {
-  try {
-    const config = JSON.parse(value.trim());
-    if (!('model_type' in config)) {
-      return 'model_type is required. Specify the model_type';
-    }
-  } catch {
-    return true;
-  }
-  return true;
-}
-
-function validateModelTypeValue(value: string) {
-  try {
-    const config = JSON.parse(value.trim());
-    if ('model_type' in config && typeof config.model_type !== 'string') {
-      return 'model_type must be a string';
-    }
-  } catch {
-    return true;
-  }
-  return true;
-}
-
-function validateEmbeddingDimensionValue(value: string) {
-  try {
-    const config = JSON.parse(value.trim());
-    if ('embedding_dimension' in config && typeof config.embedding_dimension !== 'number') {
-      return 'embedding_dimension must be a number';
-    }
-  } catch {
-    return true;
-  }
-
-  return true;
-}
-
-function validateFrameworkTypeValue(value: string) {
-  try {
-    const config = JSON.parse(value.trim());
-    if ('framework_type' in config && typeof config.framework_type !== 'string') {
-      return 'framework_type must be a string';
-    }
-  } catch {
-    return true;
-  }
-  return true;
-}
+import { ModelConfiguration } from '../common/forms/model_configuration';
 
 export const ConfigurationPanel = () => {
-  const { control } = useFormContext<ModelFileFormData | ModelUrlFormData>();
-  const [isHelpVisible, setIsHelpVisible] = useState(false);
-  const configurationFieldController = useController({
-    name: 'configuration',
-    control,
-    rules: {
-      required: { value: true, message: 'Configuration is required.' },
-      validate: {
-        [CUSTOM_FORM_ERROR_TYPES.INVALID_CONFIGURATION]: validateConfigurationObject,
-        [CUSTOM_FORM_ERROR_TYPES.CONFIGURATION_MISSING_MODEL_TYPE]: validateModelType,
-        [CUSTOM_FORM_ERROR_TYPES.INVALID_MODEL_TYPE_VALUE]: validateModelTypeValue,
-        [CUSTOM_FORM_ERROR_TYPES.INVALID_EMBEDDING_DIMENSION_VALUE]: validateEmbeddingDimensionValue,
-        [CUSTOM_FORM_ERROR_TYPES.INVALID_FRAMEWORK_TYPE_VALUE]: validateFrameworkTypeValue,
-      },
-    },
-  });
-
   return (
     <div data-test-subj="ml-registerModelConfiguration">
       <EuiText size="s">
@@ -120,39 +33,7 @@ export const ConfigurationPanel = () => {
         </small>
       </EuiText>
       <EuiSpacer size="m" />
-      <EuiFormRow
-        style={{ maxWidth: 800 }}
-        label="Configuration in JSON"
-        isInvalid={Boolean(configurationFieldController.fieldState.error)}
-        error={<ErrorMessage error={configurationFieldController.fieldState.error} />}
-        labelAppend={
-          <EuiButtonEmpty
-            onClick={() => setIsHelpVisible(true)}
-            size="xs"
-            color="primary"
-            data-test-subj="model-configuration-help-button"
-          >
-            Help
-          </EuiButtonEmpty>
-        }
-      >
-        <EuiCodeEditor
-          tabSize={2}
-          theme="sql_console"
-          width="100%"
-          showPrintMargin={false}
-          height="10rem"
-          mode="json"
-          value={configurationFieldController.field.value}
-          onChange={(value) => configurationFieldController.field.onChange(value)}
-          setOptions={{
-            fontSize: '14px',
-            enableBasicAutocompletion: true,
-            enableLiveAutocompletion: true,
-          }}
-        />
-      </EuiFormRow>
-      {isHelpVisible && <HelpFlyout onClose={() => setIsHelpVisible(false)} />}
+      <ModelConfiguration />
     </div>
   );
 };
